+++ src/main/java/nextstep/courses/domain/FreeSession.java
@@ -1,5 +1,7 @@
 package nextstep.courses.domain;
 
+import nextstep.courses.enumeration.SessionStatus;
+
 public class FreeSession extends Session {
 
     public FreeSession(Long sessionId, String title, Course course, SessionStatus sessionStatus) {
