+++ src/main/java/nextstep/courses/domain/Student.java
@@ -1,0 +1,7 @@
+package nextstep.courses.domain;
+
+import nextstep.users.domain.NsUser;
+
+public class Student {
+    NsUser nsUser;
+}
