+++ src/main/java/nextstep/courses/enumeration/SessionStatus.java
@@ -1,4 +1,4 @@
-package nextstep.courses.domain;
+package nextstep.courses.enumeration;
 
 public enum SessionStatus {
     PREPARING,
